+++ RELEASE-NOTES.md
@@ -119,6 +119,7 @@ as release notes for end user on a Recalbox upgrade.
 - Add resolution menu in Advanced Settings > Resolutions
 - Add scraper progress bar
 - Add display by filename in gamelist option
+- Add filter collections
 - Add go to game from search results and add search games from gamelist menu
 - Add retroachievements for systems PSP, Amstrad CPC, Dreamcast, Atomiswave, Naomi and Naomigd
 - Add Amiga 500 Mini controller configuration